+++ src/features/Top/Github.tsx
@@ -6,6 +6,7 @@ import SimpleBar from 'simplebar-react'
 
 import { LinkButton } from '@/components/Element/Button/LinkButton'
 
+import { formatGithubLanguages } from '@/utils/format'
 import { useFetcher } from '@/hooks/useFetcher'
 import { Repositories } from '@/types'
 
@@ -17,20 +18,7 @@ export const Github = () => {
 
   const languageProps = useMemo(() => {
     if (data) {
-      return data.user.repositories.edges.map((repo) => {
-        const totalSize = repo.node.languages.totalSize
-        return repo.node.languages.edges.flatMap((lang) => {
-          const size = parseFloat(((lang.size / totalSize) * 100).toFixed(1))
-
-          if (!size) return []
-
-          return {
-            value: parseFloat(((lang.size / totalSize) * 100).toFixed(1)),
-            color: lang.node.color,
-            name: lang.node.name,
-          }
-        })
-      })
+      return formatGithubLanguages(data)
     }
   }, [data])
